Moved 'getEnvironment' function into utils-lib file
