+++ address/get.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { failure, success } from '../libs/response-lib';
-import { getEnvironment } from "../typings/environment";
+import { getEnvironment } from '../libs/utils-lib';
 
 export const main = async event => {
   /**
