+++ user/create.js
@@ -1,5 +1,5 @@
 import * as dynamoDbLib from '../libs/dynamodb-lib';
-import { getEnvironment } from "../typings/environment";
+import { getEnvironment } from '../libs/utils-lib';
 
 export const main = async event => {
   /**
